Removed trigger:hover to compensate for latest bootstrap version.
See
https://github.com/twitter/bootstrap/commit/1fe4ed882341af29a3b709c06642e0dbb556560f#L1L70

The order in which bootstrap grabs the options for the hover has
changed. Explicitly passed options (like the trigger=hover that I'm removing
in this commit) now take precedence over `data-trigger='click'` options in the
html. So suddenly all the 'i' icons in the sidebar aren't click-to-open
anymore, but they reverted to hover.

As a stop-gap measure (read: deltaportaal site release) I thought it best to
chose to use bootstrap's default click trigger. That's what we use most of the
time now. And it is always customizable with a data-attribute now that we
don't override it with an explicit option anymore.